+++ src/StackAdmin/Resources/Commands.Resources.Test/ScenarioTests/RoleAssignmentTests.ps1
@@ -18,15 +18,55 @@ Tests retrieval of classic administrators
 #>
 function Test-RaClassicAdmins
 {
-	# Setup
-	$subscription = Get-AzureRmSubscription
+    # Setup
+    $subscription = Get-AzureRmSubscription | Where-Object { $_.Id -ieq "4004a9fd-d58e-48dc-aeb2-4a4aec58606f" }
 
-	# Test
-	$classic =  Get-AzureRmRoleAssignment -IncludeClassicAdministrators  | Where-Object { $_.Scope -ieq ('/subscriptions/' + $subscription[0].Id) -and $_.RoleDefinitionName.ToLower().Contains('administrator')}	
-	
-	# Assert
-	Assert-NotNull $classic
-	Assert-True { $classic.Length -ge 1 }
+    # Test
+    $classic =  Get-AzureRmRoleAssignment -IncludeClassicAdministrators  | Where-Object { $_.Scope -ieq ('/subscriptions/' + $subscription) -and $_.RoleDefinitionName.ToLower().Contains('administrator')}
+
+    # Assert
+    Assert-NotNull $classic
+    Assert-True { $classic.Length -ge 1 }
+}
+
+<#
+.SYNOPSIS
+Tests retrieval of classic administrators with subscription scope
+#>
+function Test-RaClassicAdminsWithScope
+{
+    $signInName = $(Get-AzureRmContext).Account.Id
+    $adminSubscription = '4004a9fd-d58e-48dc-aeb2-4a4aec58606f'
+    $nonAdminSubscription = '1de8ed49-d13d-4a32-a8cf-2aa8e944c392'
+
+    # Setup
+    Select-AzureRmSubscription -SubscriptionId $nonAdminSubscription
+
+    # Test
+    $classic = Get-AzureRmRoleAssignment -SignInName $signInName -Scope "/subscriptions/$adminSubscription" -IncludeClassicAdministrators | Where-Object { $_.RoleDefinitionName.ToLower().Contains('administrator') }
+
+    # Assert
+    Assert-NotNull $classic
+    Assert-True { $classic.Length -ge 1 }
+
+    # Setup
+    Select-AzureRmSubscription -SubscriptionId $adminSubscription
+
+    # Test
+    $classic = Get-AzureRmRoleAssignment -SignInName $signInName -Scope "/subscriptions/$adminSubscription" -IncludeClassicAdministrators | Where-Object { $_.RoleDefinitionName.ToLower().Contains('administrator') }
+
+    # Assert
+    Assert-NotNull $classic
+    Assert-True { $classic.Length -ge 1 }
+
+    # Setup
+    Select-AzureRmSubscription -SubscriptionId $adminSubscription
+
+    # Test
+    $classic = Get-AzureRmRoleAssignment -SignInName $signInName -Scope "/subscriptions/$nonAdminSubscription" -IncludeClassicAdministrators | Where-Object { $_.RoleDefinitionName.ToLower().Contains('administrator') }
+
+    # Assert
+    Assert-Null $classic
 }
 
 <#
@@ -36,7 +76,7 @@ Tests verifies negative scenarios for RoleAssignments
 function Test-RaNegativeScenarios
 {
     # Setup
-    $subscription = Get-AzureRmSubscription
+    $subscription = Get-AzureRmSubscription | Where-Object { $_.Id -ieq "4004a9fd-d58e-48dc-aeb2-4a4aec58606f" }
 
     # Bad OID returns zero role assignments
     $badOid = 'aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa'
@@ -65,7 +105,7 @@ function Test-RaDeleteByPSRoleAssignment
     # Setup
     $definitionName = 'Reader'
     $users = Get-AzureRmADUser | Select-Object -First 1 -Wait
-    $subscription = Get-AzureRmSubscription
+    $subscription = Get-AzureRmSubscription | Where-Object { $_.Id -ieq "4004a9fd-d58e-48dc-aeb2-4a4aec58606f" }
     $resourceGroups = Get-AzureRmResourceGroup | Select-Object -Last 1 -Wait
     $scope = '/subscriptions/'+ $subscription[0].Id +'/resourceGroups/' + $resourceGroups[0].ResourceGroupName
     Assert-AreEqual 1 $users.Count "There should be at least one user to run the test."
@@ -92,7 +132,7 @@ function Test-RaByScope
     # Setup
     $definitionName = 'Reader'
     $users = Get-AzureRmADUser | Select-Object -First 1 -Wait
-    $subscription = Get-AzureRmSubscription
+    $subscription = Get-AzureRmSubscription | Where-Object { $_.Id -ieq "4004a9fd-d58e-48dc-aeb2-4a4aec58606f" }
     $resourceGroups = Get-AzureRmResourceGroup | Select-Object -Last 1 -Wait
     $scope = '/subscriptions/'+ $subscription[0].Id +'/resourceGroups/' + $resourceGroups[0].ResourceGroupName
     $assignmentScope = $scope +"/"
@@ -226,7 +266,7 @@ function Test-RaValidateInputParameters ($cmdName)
     
     # Check if ResourceType is valid
     Assert-AreEqual $resource.ResourceType "Microsoft.Web/sites"
-    $subscription = Get-AzureRmSubscription | Select-Object -Last 1 -Wait
+    $subscription = Get-AzureRmSubscription | Where-Object { $_.Id -ieq "4004a9fd-d58e-48dc-aeb2-4a4aec58606f" }
     # Below invalid resource type should not return 'Not supported api version'.
     $resource.ResourceType = "Microsoft.KeyVault/"
     $invalidResourceType = "Scope '/subscriptions/"+$subscription.Id+"/resourceGroups/"+$resource.ResourceGroupName+"/providers/Microsoft.KeyVault/"+$resource.ResourceGroupName+"' should have even number of parts."
@@ -247,7 +287,7 @@ function Test-RaByServicePrincipal
     # Setup
     $definitionName = 'Contributor'
     $servicePrincipals = Get-AzureRmADServicePrincipal | Select-Object -Last 1 -Wait
-    $subscription = Get-AzureRmSubscription
+    $subscription = Get-AzureRmSubscription | Where-Object { $_.Id -ieq "4004a9fd-d58e-48dc-aeb2-4a4aec58606f" }
     $resourceGroups = Get-AzureRmResourceGroup | Select-Object -Last 1 -Wait
     $scope = '/subscriptions/'+ $subscription[0].Id +'/resourceGroups/' + $resourceGroups[0].ResourceGroupName
     Assert-AreEqual 1 $servicePrincipals.Count "No service principals found. Unable to run the test."
@@ -325,7 +365,7 @@ function Test-RaDeletionByScope
     # Setup
     $definitionName = 'Reader'
     $users = Get-AzureRmADUser | Select-Object -First 1 -Wait
-    $subscription = Get-AzureRmSubscription
+    $subscription = Get-AzureRmSubscription | Where-Object { $_.Id -ieq "4004a9fd-d58e-48dc-aeb2-4a4aec58606f" }
     $resourceGroups = Get-AzureRmResourceGroup | Select-Object -Last 1 -Wait
     $scope = '/subscriptions/'+ $subscription[0].Id +'/resourceGroups/' + $resourceGroups[0].ResourceGroupName
     Assert-AreEqual 1 $users.Count "There should be at least one user to run the test."
@@ -359,7 +399,7 @@ function Test-RaDeletionByScopeAtRootScope
     # Setup
     $definitionName = 'Reader'
     $users = Get-AzureRmADUser | Select-Object -First 1 -Wait
-    $subscription = Get-AzureRmSubscription
+    $subscription = Get-AzureRmSubscription | Where-Object { $_.Id -ieq "4004a9fd-d58e-48dc-aeb2-4a4aec58606f" }
     $resourceGroups = Get-AzureRmResourceGroup | Select-Object -Last 1 -Wait
     $scope = '/'
     Assert-AreEqual 1 $users.Count "There should be at least one user to run the test."
@@ -392,7 +432,7 @@ function Test-RaPropertiesValidation
 {
     # Setup
     $users = Get-AzureRmADUser | Select-Object -First 1 -Wait
-    $subscription = Get-AzureRmSubscription
+    $subscription = Get-AzureRmSubscription | Where-Object { $_.Id -ieq "4004a9fd-d58e-48dc-aeb2-4a4aec58606f" }
     $scope = '/subscriptions/'+$subscription[0].Id
     $roleDef = Get-AzureRmRoleDefinition -Name "Reader"
     $roleDef.Id = $null